+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - **eslint** has been upgraded from 22.0.0 to **23.0.0**
 - **express-session** has been upgraded from 1.17.0 to **1.17.2**
 - **grunt-eslint** has been removed
+- **grunt-karma** has been removed
 - **grunt-mkdocs** has been removed
 - **jsonpath-plus** has been upgraded from 3.0.0 to **6.0.1**
 - **karma** has been upgraded from 4.4.1 to **6.3.4**